Removing GymProductionServer from code
