+++ reward_kit/mcp/__init__.py
@@ -12,7 +12,6 @@ from .adapter import EnvironmentAdapter
 # New refactored components
 from .client import MCPConnectionManager
 from .execution import FireworksPolicy, LLMBasePolicy, ExecutionManager
-from .gym_production_server import GymProductionServer
 
 # North Star MCP-Gym Framework
 from .mcpgym import McpGym
@@ -26,7 +25,6 @@ __all__ = [
     "MCPEnvironmentServer",
     "EnvironmentAdapter",
     "SimulationServerBase",
-    "GymProductionServer",
     # New refactored components
     "MCPConnectionManager",
     "LLMBasePolicy",
